streamlit_app/requirements.txt
streamlit>=1.47.1
streamlit-webrtc>=0.47.0
opencv-python>=4.8.0
numpy<2
pandas>=2.0.0
plotly>=5.22.0
Pillow>=10.0.0
easyocr==1.7.1
ultralytics==8.1.0
torch>=2.0.0
scipy==1.11.4
av>=10.0.0
